moved config class to config submodule
