+++ mslib/msui/flighttrack.py
@@ -48,7 +48,7 @@ from mslib import utils, __version__
 from mslib.utils.units import units
 from mslib.utils import thermolib, find_location, save_settings_qsettings, load_settings_qsettings
 from mslib.utils.config import config_loader
-from mslib.msui import MissionSupportSystemDefaultConfig as mss_default
+from mslib.utils.config import MissionSupportSystemDefaultConfig as mss_default
 from mslib.msui.mss_qt import variant_to_string, variant_to_float
 from mslib.msui.performance_settings import DEFAULT_PERFORMANCE
